test: split PDF visual page furniture coverage
